refactor: extraer widgets del dashboard
