+++ frontend/src/components/CustomBarChart.jsx
@@ -1,0 +1,84 @@
+import React, { useMemo } from 'react';
+import { Card, CardContent, Typography, Box } from '@mui/material';
+import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts';
+
+const CustomBarChart = React.memo(({ data, xKey, barKey, title, isDarkMode, height = 300 }) => {
+    const chartData = useMemo(() => data, [data]);
+    return (
+        <Card sx={{
+            height: '100%',
+            background: isDarkMode
+                ? 'rgba(45, 55, 72, 0.8)'
+                : 'rgba(255, 255, 255, 0.9)',
+            backdropFilter: 'blur(20px)',
+            border: isDarkMode
+                ? '1px solid rgba(255, 255, 255, 0.1)'
+                : '1px solid rgba(0, 0, 0, 0.08)',
+            borderRadius: 3,
+            transition: 'all 0.3s ease',
+            '&:hover': {
+                transform: 'translateY(-4px)',
+                boxShadow: isDarkMode
+                    ? '0 12px 40px rgba(0, 0, 0, 0.4)'
+                    : '0 12px 40px rgba(0, 0, 0, 0.15)',
+            }
+        }}>
+            <CardContent sx={{ p: 3 }}>
+                <Typography
+                    variant="h6"
+                    gutterBottom
+                    align="center"
+                    sx={{
+                        fontWeight: 600,
+                        color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
+                        mb: 2,
+                    }}
+                >
+                    {title}
+                </Typography>
+                <Box sx={{ height: height }}>
+                    <ResponsiveContainer width="100%" height="100%">
+                        <BarChart data={chartData} margin={{ top: 5, right: 20, left: -10, bottom: 80 }}>
+                            <CartesianGrid
+                                strokeDasharray="3 3"
+                                stroke={isDarkMode ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.1)'}
+                            />
+                            <XAxis
+                                dataKey={xKey}
+                                tick={{
+                                    fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)',
+                                    fontSize: xKey === 'range' ? 14 : 10
+                                }}
+                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
+                                angle={xKey === 'range' ? 0 : -45}
+                                textAnchor={xKey === 'range' ? 'middle' : 'end'}
+                                height={80}
+                                interval={0}
+                            />
+                            <YAxis
+                                tick={{ fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)' }}
+                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
+                            />
+                            <Tooltip
+                                contentStyle={{
+                                    backgroundColor: isDarkMode ? 'rgba(45, 55, 72, 0.95)' : 'rgba(255, 255, 255, 0.95)',
+                                    border: isDarkMode ? '1px solid rgba(255, 255, 255, 0.1)' : '1px solid rgba(0, 0, 0, 0.1)',
+                                    borderRadius: '8px',
+                                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
+                                }}
+                                labelFormatter={(label) => `${xKey === 'function' ? 'Función' : xKey === 'range' ? 'Rango de edad' : xKey === 'area' ? 'Área' : 'Categoría'}: ${label}`}
+                                formatter={(value, name) => [
+                                    barKey === 'avgAge' ? `${Math.round(value)} años` : value,
+                                    barKey === 'avgAge' ? 'Edad promedio' : 'Cantidad de agentes'
+                                ]}
+                            />
+                            <Bar dataKey={barKey} fill="#00C49F" radius={[4, 4, 0, 0]} />
+                        </BarChart>
+                    </ResponsiveContainer>
+                </Box>
+            </CardContent>
+        </Card>
+    );
+});
+
+export default CustomBarChart;
