+++ frontend/src/components/CustomDonutChart.jsx
@@ -1,0 +1,133 @@
+import React, { useMemo } from 'react';
+import { Card, CardContent, Typography, Box } from '@mui/material';
+import { PieChart, Pie, Cell, Tooltip, Legend, ResponsiveContainer } from 'recharts';
+
+const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042', '#8884D8', '#82CA9D', '#FFC658', '#FF7C7C', '#8DD1E1', '#D084D0'];
+
+const CustomDonutChart = React.memo(({ data, title, isDarkMode, dataKey, nameKey, height = 400 }) => {
+    const chartData = useMemo(() => data, [data]);
+
+    const total = useMemo(() => {
+        return chartData.reduce((sum, item) => sum + (item[dataKey] || 0), 0);
+    }, [chartData, dataKey]);
+
+    const CustomTooltip = ({ active, payload }) => {
+        if (active && payload && payload.length) {
+            const data = payload[0].payload;
+            const percentage = total > 0 ? ((data[dataKey] / total) * 100).toFixed(1) : 0;
+            return (
+                <Box sx={{
+                    backgroundColor: isDarkMode ? 'rgba(45, 55, 72, 0.95)' : 'rgba(255, 255, 255, 0.95)',
+                    border: isDarkMode ? '1px solid rgba(255, 255, 255, 0.1)' : '1px solid rgba(0, 0, 0, 0.1)',
+                    borderRadius: '8px',
+                    p: 2,
+                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
+                }}>
+                    <Typography variant="body2" sx={{ fontWeight: 600 }}>
+                        {data[nameKey]}
+                    </Typography>
+                    <Typography variant="body2">
+                        Cantidad de agentes: {data[dataKey]}
+                    </Typography>
+                    <Typography variant="body2">
+                        Porcentaje: {percentage}%
+                    </Typography>
+                </Box>
+            );
+        }
+        return null;
+    };
+
+    const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, percent }) => {
+        if (percent < 0.05) return null;
+
+        const RADIAN = Math.PI / 180;
+        const radius = innerRadius + (outerRadius - innerRadius) * 0.5;
+        const x = cx + radius * Math.cos(-midAngle * RADIAN);
+        const y = cy + radius * Math.sin(-midAngle * RADIAN);
+
+        return (
+            <text
+                x={x}
+                y={y}
+                fill={isDarkMode ? 'white' : 'black'}
+                textAnchor={x > cx ? 'start' : 'end'}
+                dominantBaseline="central"
+                fontSize="12"
+                fontWeight="600"
+            >
+                {`${(percent * 100).toFixed(0)}%`}
+            </text>
+        );
+    };
+
+    return (
+        <Card sx={{
+            height: height,
+            background: isDarkMode
+                ? 'rgba(45, 55, 72, 0.8)'
+                : 'rgba(255, 255, 255, 0.9)',
+            backdropFilter: 'blur(20px)',
+            border: isDarkMode
+                ? '1px solid rgba(255, 255, 255, 0.1)'
+                : '1px solid rgba(0, 0, 0, 0.08)',
+            borderRadius: 3,
+            transition: 'all 0.3s ease',
+            '&:hover': {
+                transform: 'translateY(-4px)',
+                boxShadow: isDarkMode
+                    ? '0 12px 40px rgba(0, 0, 0, 0.4)'
+                    : '0 12px 40px rgba(0, 0, 0, 0.15)',
+            }
+        }}>
+            <CardContent sx={{ p: 3, height: '100%', display: 'flex', flexDirection: 'column' }}>
+                <Typography
+                    variant="h6"
+                    gutterBottom
+                    align="center"
+                    sx={{
+                        fontWeight: 600,
+                        color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
+                        mb: 2,
+                    }}
+                >
+                    {title}
+                </Typography>
+                <Box sx={{ flexGrow: 1, minHeight: 300 }}>
+                    <ResponsiveContainer width="100%" height="100%">
+                        <PieChart>
+                            <Pie
+                                data={chartData}
+                                cx="50%"
+                                cy="50%"
+                                labelLine={false}
+                                label={renderCustomizedLabel}
+                                outerRadius={100}
+                                innerRadius={60}
+                                fill="#8884d8"
+                                dataKey={dataKey}
+                                nameKey={nameKey}
+                            >
+                                {chartData.map((entry, index) => (
+                                    <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
+                                ))}
+                            </Pie>
+                            <Tooltip content={<CustomTooltip />} />
+                            <Legend
+                                wrapperStyle={{
+                                    color: isDarkMode ? 'rgba(255, 255, 255, 0.8)' : 'rgba(0, 0, 0, 0.7)',
+                                    fontSize: '12px',
+                                    paddingTop: '10px'
+                                }}
+                                iconSize={8}
+                                formatter={(value) => value}
+                            />
+                        </PieChart>
+                    </ResponsiveContainer>
+                </Box>
+            </CardContent>
+        </Card>
+    );
+});
+
+export default CustomDonutChart;
