+++ frontend/src/page/DashboardPage.jsx
@@ -1,398 +1,17 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useAuth } from '../context/AuthContext';
-import { Box, Typography, Card, CardContent, CircularProgress, Alert, Grid, Button } from '@mui/material';
+import { Box, Typography, CircularProgress, Alert, Grid, Button } from '@mui/material';
 import { useTheme } from '../context/ThemeContext.jsx';
-import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer, PieChart, Pie, Cell, AreaChart, Area } from 'recharts';
 import apiClient from '../services/api';
 import DashboardIcon from '@mui/icons-material/Dashboard';
 import AnalyticsIcon from '@mui/icons-material/Analytics';
 import BusinessIcon from '@mui/icons-material/Business';
 import CleaningServicesIcon from '@mui/icons-material/CleaningServices';
+import StatCard from '../components/StatCard';
+import CustomBarChart from '../components/CustomBarChart';
+import CustomDonutChart from '../components/CustomDonutChart';
+import CustomAreaChart from '../components/CustomAreaChart';
 
-const StatCard = React.memo(({ title, value, color = 'primary.main', isDarkMode }) => (
-    <Card sx={{ 
-        height: '100%',
-        background: isDarkMode
-            ? 'rgba(45, 55, 72, 0.8)'
-            : 'rgba(255, 255, 255, 0.9)',
-        backdropFilter: 'blur(20px)',
-        border: isDarkMode
-            ? '1px solid rgba(255, 255, 255, 0.1)'
-            : '1px solid rgba(0, 0, 0, 0.08)',
-        borderRadius: 3,
-        transition: 'all 0.3s ease',
-        '&:hover': {
-            transform: 'translateY(-4px)',
-            boxShadow: isDarkMode
-                ? '0 12px 40px rgba(0, 0, 0, 0.4)'
-                : '0 12px 40px rgba(0, 0, 0, 0.15)',
-        }
-    }}>
-        <CardContent sx={{ p: 3 }}>
-            <Typography 
-                color="text.secondary" 
-                gutterBottom
-                sx={{
-                    fontWeight: 600,
-                    fontSize: '0.9rem',
-                    color: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.6)',
-                }}
-            >
-                {title}
-            </Typography>
-            <Typography 
-                variant="h4" 
-                component="div" 
-                sx={{ 
-                    fontWeight: 700,
-                    fontSize: '2rem',
-                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                }}
-            >
-                {value}
-            </Typography>
-        </CardContent>
-    </Card>
-));
-
-const CustomBarChart = React.memo(({ data, xKey, barKey, title, isDarkMode, height = 300 }) => {
-    const chartData = useMemo(() => data, [data]);
-    return (
-        <Card sx={{ 
-            height: '100%',
-            background: isDarkMode
-                ? 'rgba(45, 55, 72, 0.8)'
-                : 'rgba(255, 255, 255, 0.9)',
-            backdropFilter: 'blur(20px)',
-            border: isDarkMode
-                ? '1px solid rgba(255, 255, 255, 0.1)'
-                : '1px solid rgba(0, 0, 0, 0.08)',
-            borderRadius: 3,
-            transition: 'all 0.3s ease',
-            '&:hover': {
-                transform: 'translateY(-4px)',
-                boxShadow: isDarkMode
-                    ? '0 12px 40px rgba(0, 0, 0, 0.4)'
-                    : '0 12px 40px rgba(0, 0, 0, 0.15)',
-            }
-        }}>
-            <CardContent sx={{ p: 3 }}>
-                <Typography 
-                    variant="h6" 
-                    gutterBottom 
-                    align="center"
-                    sx={{
-                        fontWeight: 600,
-                        color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                        mb: 2,
-                    }}
-                >
-                    {title}
-                </Typography>
-                <Box sx={{ height: height }}>
-                    <ResponsiveContainer width="100%" height="100%">
-                        <BarChart data={chartData} margin={{ top: 5, right: 20, left: -10, bottom: 80 }}>
-                            <CartesianGrid 
-                                strokeDasharray="3 3" 
-                                stroke={isDarkMode ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.1)'}
-                            />
-                            <XAxis 
-                                dataKey={xKey} 
-                                tick={{ 
-                                    fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)',
-                                    fontSize: xKey === 'range' ? 14 : 10
-                                }}
-                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
-                                angle={xKey === 'range' ? 0 : -45}
-                                textAnchor={xKey === 'range' ? 'middle' : 'end'}
-                                height={80}
-                                interval={0}
-                            />
-                            <YAxis 
-                                tick={{ fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)' }}
-                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
-                            />
-                            <Tooltip 
-                                contentStyle={{
-                                    backgroundColor: isDarkMode ? 'rgba(45, 55, 72, 0.95)' : 'rgba(255, 255, 255, 0.95)',
-                                    border: isDarkMode ? '1px solid rgba(255, 255, 255, 0.1)' : '1px solid rgba(0, 0, 0, 0.1)',
-                                    borderRadius: '8px',
-                                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                                }}
-                                labelFormatter={(label) => `${xKey === 'function' ? 'Función' : xKey === 'range' ? 'Rango de edad' : xKey === 'area' ? 'Área' : 'Categoría'}: ${label}`}
-                                formatter={(value, name) => [
-                                    barKey === 'avgAge' ? `${Math.round(value)} años` : value, 
-                                    barKey === 'avgAge' ? 'Edad promedio' : 'Cantidad de agentes'
-                                ]}
-                            />
-                            <Bar dataKey={barKey} fill="#00C49F" radius={[4, 4, 0, 0]} />
-                        </BarChart>
-                    </ResponsiveContainer>
-                </Box>
-            </CardContent>
-        </Card>
-    );
-});
-
-// Colores para los gráficos de torta
-const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042', '#8884D8', '#82CA9D', '#FFC658', '#FF7C7C', '#8DD1E1', '#D084D0'];
-
-// Componente de gráfico de anillo personalizado UNIFICADO CON LEGEND
-const CustomDonutChartUnified = React.memo(({ data, title, isDarkMode, dataKey, nameKey, height = 400 }) => {
-    const chartData = useMemo(() => data, [data]);
-    
-    const total = useMemo(() => {
-        return chartData.reduce((sum, item) => sum + (item[dataKey] || 0), 0);
-    }, [chartData, dataKey]);
-    
-    const CustomTooltip = ({ active, payload }) => {
-        if (active && payload && payload.length) {
-            const data = payload[0].payload;
-            const percentage = total > 0 ? ((data[dataKey] / total) * 100).toFixed(1) : 0;
-            return (
-                <Box sx={{
-                    backgroundColor: isDarkMode ? 'rgba(45, 55, 72, 0.95)' : 'rgba(255, 255, 255, 0.95)',
-                    border: isDarkMode ? '1px solid rgba(255, 255, 255, 0.1)' : '1px solid rgba(0, 0, 0, 0.1)',
-                    borderRadius: '8px',
-                    p: 2,
-                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                }}>
-                    <Typography variant="body2" sx={{ fontWeight: 600 }}>
-                        {data[nameKey]}
-                    </Typography>
-                    <Typography variant="body2">
-                        Cantidad de agentes: {data[dataKey]}
-                    </Typography>
-                    <Typography variant="body2">
-                        Porcentaje: {percentage}%
-                    </Typography>
-                </Box>
-            );
-        }
-        return null;
-    };
-    
-    const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, percent }) => {
-        if (percent < 0.05) return null; // No mostrar etiquetas para segmentos muy pequeños
-        
-        const RADIAN = Math.PI / 180;
-        const radius = innerRadius + (outerRadius - innerRadius) * 0.5;
-        const x = cx + radius * Math.cos(-midAngle * RADIAN);
-        const y = cy + radius * Math.sin(-midAngle * RADIAN);
-
-        return (
-            <text 
-                x={x} 
-                y={y} 
-                fill={isDarkMode ? 'white' : 'black'} 
-                textAnchor={x > cx ? 'start' : 'end'} 
-                dominantBaseline="central"
-                fontSize="12"
-                fontWeight="600"
-            >
-                {`${(percent * 100).toFixed(0)}%`}
-            </text>
-        );
-    };
-    
-    return (
-        <Card sx={{ 
-            height: height, // Altura personalizable
-            background: isDarkMode
-                ? 'rgba(45, 55, 72, 0.8)'
-                : 'rgba(255, 255, 255, 0.9)',
-            backdropFilter: 'blur(20px)',
-            border: isDarkMode
-                ? '1px solid rgba(255, 255, 255, 0.1)'
-                : '1px solid rgba(0, 0, 0, 0.08)',
-            borderRadius: 3,
-            transition: 'all 0.3s ease',
-            '&:hover': {
-                transform: 'translateY(-4px)',
-                boxShadow: isDarkMode
-                    ? '0 12px 40px rgba(0, 0, 0, 0.4)'
-                    : '0 12px 40px rgba(0, 0, 0, 0.15)',
-            }
-        }}>
-            <CardContent sx={{ p: 3, height: '100%', display: 'flex', flexDirection: 'column' }}>
-                <Typography 
-                    variant="h6" 
-                    gutterBottom 
-                    align="center"
-                    sx={{
-                        fontWeight: 600,
-                        color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                        mb: 2,
-                    }}
-                >
-                    {title}
-                </Typography>
-                <Box sx={{ 
-                    display: 'flex', 
-                    justifyContent: 'center', 
-                    alignItems: 'center',
-                    mb: 2 
-                }}>
-                    <Typography 
-                        variant="h4" 
-                        sx={{ 
-                            fontWeight: 700,
-                            color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                        }}
-                    >
-                        {total.toLocaleString()}
-                    </Typography>
-                    <Typography 
-                        variant="body2" 
-                        sx={{ 
-                            ml: 1,
-                            color: isDarkMode ? 'rgba(255, 255, 255, 0.6)' : 'rgba(0, 0, 0, 0.6)',
-                        }}
-                    >
-                        Total
-                    </Typography>
-                </Box>
-                <Box sx={{ flexGrow: 1, minHeight: 300 }}>
-                    <ResponsiveContainer width="100%" height="100%">
-                        <PieChart>
-                            <Pie
-                                data={chartData}
-                                cx="50%"
-                                cy="50%"
-                                labelLine={false}
-                                label={renderCustomizedLabel}
-                                outerRadius={100}
-                                innerRadius={60}
-                                fill="#8884d8"
-                                dataKey={dataKey}
-                                nameKey={nameKey}
-                            >
-                                {chartData.map((entry, index) => (
-                                    <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                                ))}
-                            </Pie>
-                            <Tooltip content={<CustomTooltip />} />
-                            <Legend 
-                                wrapperStyle={{
-                                    color: isDarkMode ? 'rgba(255, 255, 255, 0.8)' : 'rgba(0, 0, 0, 0.7)',
-                                    fontSize: '12px',
-                                    paddingTop: '10px'
-                                }}
-                                iconSize={8}
-                                formatter={(value) => value} // Mostrar el nombre real, no "count"
-                            />
-                        </PieChart>
-                    </ResponsiveContainer>
-                </Box>
-            </CardContent>
-        </Card>
-    );
-});
-
-// Componente de gráfico de área personalizado para análisis por área
-const CustomAreaChartByArea = React.memo(({ data, title, isDarkMode, xKey, yKey }) => {
-    const chartData = useMemo(() => data, [data]);
-    
-    const CustomTooltip = ({ active, payload, label }) => {
-        if (active && payload && payload.length) {
-            return (
-                <Box sx={{
-                    backgroundColor: isDarkMode ? 'rgba(45, 55, 72, 0.95)' : 'rgba(255, 255, 255, 0.95)',
-                    border: isDarkMode ? '1px solid rgba(255, 255, 255, 0.1)' : '1px solid rgba(0, 0, 0, 0.1)',
-                    borderRadius: '8px',
-                    p: 2,
-                    color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                }}>
-                    <Typography variant="body2" sx={{ fontWeight: 600 }}>
-                        Área: {label}
-                    </Typography>
-                    <Typography variant="body2">
-                        Edad promedio: {Math.round(payload[0].value)} años
-                    </Typography>
-                </Box>
-            );
-        }
-        return null;
-    };
-    
-    return (
-        <Card sx={{ 
-            height: '100%',
-            background: isDarkMode
-                ? 'rgba(45, 55, 72, 0.8)'
-                : 'rgba(255, 255, 255, 0.9)',
-            backdropFilter: 'blur(20px)',
-            border: isDarkMode
-                ? '1px solid rgba(255, 255, 255, 0.1)'
-                : '1px solid rgba(0, 0, 0, 0.08)',
-            borderRadius: 3,
-            transition: 'all 0.3s ease',
-            '&:hover': {
-                transform: 'translateY(-4px)',
-                boxShadow: isDarkMode
-                    ? '0 12px 40px rgba(0, 0, 0, 0.4)'
-                    : '0 12px 40px rgba(0, 0, 0, 0.15)',
-            }
-        }}>
-            <CardContent sx={{ p: 3 }}>
-                <Typography 
-                    variant="h6" 
-                    gutterBottom 
-                    align="center"
-                    sx={{
-                        fontWeight: 600,
-                        color: isDarkMode ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)',
-                        mb: 2,
-                    }}
-                >
-                    {title}
-                </Typography>
-                <Box sx={{ height: 300 }}>
-                    <ResponsiveContainer width="100%" height="100%">
-                        <AreaChart data={chartData} margin={{ top: 10, right: 30, left: 0, bottom: 80 }}>
-                            <defs>
-                                <linearGradient id="colorAreaByArea" x1="0" y1="0" x2="0" y2="1">
-                                    <stop offset="5%" stopColor="#00C49F" stopOpacity={0.8}/>
-                                    <stop offset="95%" stopColor="#00C49F" stopOpacity={0.1}/>
-                                </linearGradient>
-                            </defs>
-                            <CartesianGrid 
-                                strokeDasharray="3 3" 
-                                stroke={isDarkMode ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.1)'}
-                            />
-                            <XAxis 
-                                dataKey={xKey}
-                                tick={{ 
-                                    fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)',
-                                    fontSize: 10
-                                }}
-                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
-                                angle={-45}
-                                textAnchor="end"
-                                height={80}
-                                interval={0}
-                            />
-                            <YAxis 
-                                tick={{ fill: isDarkMode ? 'rgba(255, 255, 255, 0.7)' : 'rgba(0, 0, 0, 0.7)' }}
-                                axisLine={{ stroke: isDarkMode ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.3)' }}
-                            />
-                            <Tooltip content={<CustomTooltip />} />
-                            <Area 
-                                type="monotone" 
-                                dataKey={yKey} 
-                                stroke="#00C49F" 
-                                fillOpacity={1} 
-                                fill="url(#colorAreaByArea)" 
-                                strokeWidth={2}
-                                name="Edad promedio"
-                            />
-                        </AreaChart>
-                    </ResponsiveContainer>
-                </Box>
-            </CardContent>
-        </Card>
-    );
-});
 
 // Función para hacer peticiones con reintentos
 const fetchWithRetry = async (url, maxRetries = 3, delay = 1000) => {
@@ -857,7 +476,7 @@ const DashboardPage = () => {
 
                     {/* Gráficos principales - AMBOS CON MISMA ALTURA */}
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByFunction.filter(f => f.function && f.function.trim() !== '' && f.function.trim() !== '-').slice(0, 10)} 
                             title="Distribución de Agentes por Función (Top 10) - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -867,7 +486,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByEmploymentType} 
                             title="Agentes por Situación de Revista - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -878,7 +497,7 @@ const DashboardPage = () => {
                     </Grid>
                     {/* NUEVOS GRÁFICOS PARA NEIKES Y BECA */}
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByFunctionNeikeBeca.filter(f => f.function && f.function.trim() !== '' && f.function.trim() !== '-').slice(0, 10)} 
                             title="Distribución de Agentes por Función (Top 10) - Neikes y Beca" 
                             isDarkMode={isDarkMode}
@@ -888,7 +507,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByEmploymentTypeNeikeBeca} 
                             title="Agentes por Situación de Revista - Neikes y Becas" 
                             isDarkMode={isDarkMode}
@@ -953,7 +572,7 @@ const DashboardPage = () => {
                     {/* Edad promedio por área (secretaría) - PLANTA Y CONTRATOS */}
                     <Grid item xs={12} lg={6}>
                         {ageByArea.length > 0 ? (
-                            <CustomAreaChartByArea 
+                            <CustomAreaChart 
                                 data={ageByArea.filter(a => a.secretaria && a.secretaria.trim() !== '' && a.secretaria.trim() !== '-').slice(0, 10)} 
                                 title="Distribución por rangos de edad según el área (Top 10) - Planta y Contratos" 
                                 isDarkMode={isDarkMode}
@@ -988,7 +607,7 @@ const DashboardPage = () => {
                     {/* Edad promedio por área (secretaría) - NEIKES Y BECAS */}
                     <Grid item xs={12} lg={6}>
                         {ageByAreaNeikeBeca.length > 0 ? (
-                            <CustomAreaChartByArea 
+                            <CustomAreaChart 
                                 data={ageByAreaNeikeBeca.filter(a => a.secretaria && a.secretaria.trim() !== '' && a.secretaria.trim() !== '-').slice(0, 10)} 
                                 title="Distribución por rangos de edad según el área (Top 10) - Neikes y Becas" 
                                 isDarkMode={isDarkMode}
@@ -1036,7 +655,7 @@ const DashboardPage = () => {
 
                     {/* Gráficos de anillo principales */}
                     <Grid item xs={12} md={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsBySecretaria.slice(0, 8)} 
                             title="Agentes por Secretaría (Top 8) - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -1046,7 +665,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} md={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByDependency.slice(0, 8)} 
                             title="Agentes por Dependencia (Top 8) - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -1077,7 +696,7 @@ const DashboardPage = () => {
 
                     {/* Gráficos de anillo principales - Neikes y Becas */}
                     <Grid item xs={12} md={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsBySecretariaNeikeBeca ? agentsBySecretariaNeikeBeca.slice(0, 8) : []} 
                             title="Agentes por Secretaría (Top 8) - Neikes y Becas" 
                             isDarkMode={isDarkMode}
@@ -1087,7 +706,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} md={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByDependencyNeikeBeca ? agentsByDependencyNeikeBeca.slice(0, 8) : []} 
                             title="Agentes por Dependencia (Top 8) - Neikes y Becas" 
                             isDarkMode={isDarkMode}
@@ -1176,7 +795,7 @@ const DashboardPage = () => {
 
                     {/* Departamentos y Divisiones */}
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={filterValidData(agentsByDepartamento, 'departamento').slice(0, 8)} 
                             title="Agentes por Departamento (Top 8) - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -1186,7 +805,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={filterValidData(agentsByDivision, 'division').slice(0, 8)} 
                             title="Agentes por División (Top 8) - Planta y Contratos" 
                             isDarkMode={isDarkMode}
@@ -1205,7 +824,7 @@ const DashboardPage = () => {
 
                     {/* Departamentos y Divisiones - Neikes y Becas */}
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByDepartamentoNeikeBeca ? filterValidData(agentsByDepartamentoNeikeBeca, 'departamento').slice(0, 8) : []} 
                             title="Agentes por Departamento (Top 8) - Neikes y Becas" 
                             isDarkMode={isDarkMode}
@@ -1215,7 +834,7 @@ const DashboardPage = () => {
                         />
                     </Grid>
                     <Grid item xs={12} lg={6}>
-                        <CustomDonutChartUnified 
+                        <CustomDonutChart 
                             data={agentsByDivisionNeikeBeca ? filterValidData(agentsByDivisionNeikeBeca, 'division').slice(0, 8) : []} 
                             title="Agentes por División (Top 8) - Neikes y Becas" 
                             isDarkMode={isDarkMode}
